# docs/swagger.yaml
basePath: /
definitions:
  admin.BanUserRequest:
    properties:
      reason:
        type: string
    type: object
  admin.ProductRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  admin.UpdateRoleRequest:
    properties:
      role:
        type: string
    type: object
...
  reqresp.ErrorInfo:
    properties:
      code:
        example: 400
        type: integer
        x-order: "1"
      details:
        example: 'Invalid ID format: expected integer, got ''abc'''
        type: string
        x-order: "2"
    type: object
...
  reqresp.StandardResponse:
    properties:
      data:
        x-order: "3"
      error:
        allOf:
        - $ref: '#/definitions/reqresp.ErrorInfo'
        x-order: "4"
      message:
        example: Operation completed successfully
        type: string
        x-order: "2"
      success:
        example: true
        type: boolean
        x-order: "1"
    type: object
...
info:
  contact: {}
  description: API documentation for Marketplace project
  title: Go Marketplace API
  version: "1.0"
paths:
  /api/admin/dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get admin dashboard statistics
      tags:
      - admin
  /api/admin/products:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
      security:
      - BearerAuth: []
      summary: List all products
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Product data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/admin.ProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
      security:
      - BearerAuth: []
      summary: Create a product
      tags:
      - admin
  /api/admin/products/{id}:
    delete:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - admin
    get:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get product by ID
      tags:
      - admin
    put:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/admin.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - admin
  /api/admin/stats/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get user statistics
      tags:
      - admin
  /api/admin/users:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by role (admin, seller, customer)
        in: query
        name: role
        type: string
      - description: Filter by ban status (true, false)
        in: query
        name: is_banned
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - admin
  /api/admin/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - admin
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - admin
  /api/admin/users/{id}/ban:
    post:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ban reason
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/admin.BanUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
      security:
      - BearerAuth: []
      summary: Ban a user
      tags:
      - admin
  /api/admin/users/{id}/role:
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: New role
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
      security:
      - BearerAuth: []
      summary: Update user role
      tags:
      - admin
  /api/admin/users/{id}/unban:
    post:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqresp.StandardResponse'
      security:
      - BearerAuth: []
      summary: Unban a user
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"